requirements.txt
django
djangorestframework
drf-spectacular
ruff
python-decouple
pytest
pytest-django
celery
redis
django-redis
gunicorn
psycopg2-binary
sqlalchemy
django-debug-toolbar
uuid
pillow
faker
djangorestframework-simplejwt
django-ratelimit
cryptography